requirements.txt
# Core
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Auth
pyjwt>=2.8.0

# FEA / Numerics
numpy>=1.26.0
scipy>=1.12.0
matplotlib>=3.8.0

# Testing / Dev
pytest>=8.0.0,<9.0.0
httpx>=0.27.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
numpy>=2.0.0,<3.0.0
scipy>=1.13.0,<2.0.0
